# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeInput {
  /** YouTube URL (video or playlist) */
  url: string;
}

/**
 * Download format
 */
export type PlaylistDownloadInputFormat = typeof PlaylistDownloadInputFormat[keyof typeof PlaylistDownloadInputFormat];


export const PlaylistDownloadInputFormat = {
  video: 'video',
  audio: 'audio',
} as const;

/**
 * Video quality
 */
export type PlaylistDownloadInputQuality = typeof PlaylistDownloadInputQuality[keyof typeof PlaylistDownloadInputQuality];


export const PlaylistDownloadInputQuality = {
  highest: 'highest',
  lowest: 'lowest',
  '1080p': '1080p',
  '720p': '720p',
  '480p': '480p',
  '360p': '360p',
  '240p': '240p',
  '144p': '144p',
} as const;

export interface PlaylistDownloadInput {
  /** YouTube video URL within the playlist */
  url: string;
  /** Download format */
  format: PlaylistDownloadInputFormat;
  /** Video quality */
  quality?: PlaylistDownloadInputQuality;
}

export type VideoInfoType = typeof VideoInfoType[keyof typeof VideoInfoType];


export const VideoInfoType = {
  video: 'video',
} as const;

export interface VideoInfo {
  type: VideoInfoType;
  videoId: string;
  title: string;
  thumbnail: string;
  duration: string;
  durationSeconds: number;
  channel: string;
  /** @nullable */
  channelUrl: string | null;
  /** @nullable */
  channelAvatar?: string | null;
  /** @nullable */
  subscribers?: string | null;
  /** @nullable */
  viewCount: number | null;
  /** @nullable */
  uploadDate: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  category?: string | null;
  isShort?: boolean;
  url: string;
}

export interface PlaylistItem {
  videoId: string;
  title: string;
  thumbnail: string;
  duration: string;
  durationSeconds: number;
  /** @nullable */
  channel?: string | null;
  url: string;
}

export type PlaylistInfoType = typeof PlaylistInfoType[keyof typeof PlaylistInfoType];


export const PlaylistInfoType = {
  playlist: 'playlist',
} as const;

export interface PlaylistInfo {
  type: PlaylistInfoType;
  playlistId: string;
  title: string;
  thumbnail: string;
  videoCount: number;
  totalDuration: string;
  totalDurationSeconds?: number;
  /** @nullable */
  creator: string | null;
  url: string;
  videos: PlaylistItem[];
}

export type AnalyzeResult = VideoInfo | PlaylistInfo;

export interface ApiError {
  error: string;
  message: string;
  /** @nullable */
  code?: string | null;
}

export type DownloadMediaParams = {
/**
 * YouTube video URL
 */
url: string;
/**
 * Download format (video = mp4, audio = mp3)
 */
format: DownloadMediaFormat;
/**
 * Video quality (only applies to video format)
 */
quality?: DownloadMediaQuality;
};

export type DownloadMediaFormat = typeof DownloadMediaFormat[keyof typeof DownloadMediaFormat];


export const DownloadMediaFormat = {
  video: 'video',
  audio: 'audio',
} as const;

export type DownloadMediaQuality = typeof DownloadMediaQuality[keyof typeof DownloadMediaQuality];


export const DownloadMediaQuality = {
  highest: 'highest',
  lowest: 'lowest',
  '1080p': '1080p',
  '720p': '720p',
  '480p': '480p',
  '360p': '360p',
  '240p': '240p',
  '144p': '144p',
} as const;