# lib/api-zod/src/generated/types/analyzeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeInput {
  /** YouTube URL (video or playlist) */
  url: string;
}