# lib/api-zod/src/generated/types/analyzeResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistInfo } from './playlistInfo';
import type { VideoInfo } from './videoInfo';

export type AnalyzeResult = VideoInfo | PlaylistInfo;