# lib/api-zod/src/generated/types/playlistDownloadInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistDownloadInputFormat } from './playlistDownloadInputFormat';
import type { PlaylistDownloadInputQuality } from './playlistDownloadInputQuality';

export interface PlaylistDownloadInput {
  /** YouTube video URL within the playlist */
  url: string;
  /** Download format */
  format: PlaylistDownloadInputFormat;
  /** Video quality */
  quality?: PlaylistDownloadInputQuality;
}