# lib/api-zod/src/generated/types/playlistInfo.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlaylistInfoType } from './playlistInfoType';
import type { PlaylistItem } from './playlistItem';

export interface PlaylistInfo {
  type: PlaylistInfoType;
  playlistId: string;
  title: string;
  thumbnail: string;
  videoCount: number;
  totalDuration: string;
  totalDurationSeconds?: number;
  /** @nullable */
  creator: string | null;
  url: string;
  videos: PlaylistItem[];
}