# lib/api-zod/src/generated/types/playlistItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlaylistItem {
  videoId: string;
  title: string;
  thumbnail: string;
  duration: string;
  durationSeconds: number;
  /** @nullable */
  channel?: string | null;
  url: string;
}