# lib/api-zod/src/generated/types/videoInfo.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VideoInfoType } from './videoInfoType';

export interface VideoInfo {
  type: VideoInfoType;
  videoId: string;
  title: string;
  thumbnail: string;
  duration: string;
  durationSeconds: number;
  channel: string;
  /** @nullable */
  channelUrl: string | null;
  /** @nullable */
  channelAvatar?: string | null;
  /** @nullable */
  subscribers?: string | null;
  /** @nullable */
  viewCount: number | null;
  /** @nullable */
  uploadDate: string | null;
  /** @nullable */
  description: string | null;
  /** @nullable */
  category?: string | null;
  isShort?: boolean;
  url: string;
}